pyproject.toml
[tool.poetry]
name = "nonebot-plugin-calc-game"
version = "1.0.0"
description = "这是利用机器人复刻 计算器：游戏 的一个插件。通过给出的操作方式由当前数字达到计算目标。"
authors = ["Yurchiu"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Yurchiu/nonebot-plugin-calc-game"
repository = "https://github.com/Yurchiu/nonebot-plugin-calc-game"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = ">=2.3.2"
nonebot-plugin-txt2img = ">=0.4.1"
nonebot_plugin_alconna = ">=0.55.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
